Reorganize directory structure
- Move patterns/ to internal/patterns/ since it's only used internally
- Move CHANGELOG.md and IMPROVEMENTS.md to docs/
- Remove orphaned config/config.toml (embedded config is in internal/config/)
- Update import paths accordingly
