+++ internal/patterns/patterns.go
