+++ internal/patterns/patterns_test.go
